+++ lib/amq/protocol.rb
@@ -2,3 +2,4 @@
 
 require "amq/protocol/version"
 require "amq/protocol/client"
+require "amq/protocol/channel_close"